More timing improvements and server work

+++ platform/macosx/au.h
@@ -17,4 +17,4 @@ void au_send_midi(au_context_t context, int status, int data1, int data2);
 void au_render(au_context_t context, double time, int count, double *output);
 void au_cleanup(au_context_t context);
 
-void fa_midi_message_decons(fa_midi_message_t midi_message, int *statusCh, int *data1, int *data2);
+void fa_midi_message_decons(fa_midi_message_t midi_message, uint8_t *statusCh, uint8_t *data1, uint8_t *data2);
