+++ platform/macosx/signal.c
@@ -82,7 +82,7 @@ fa_ptr_t receive_(fa_ptr_t x, fa_signal_name_t n, fa_signal_message_t msg)
         if (!fa_midi_message_is_simple(msg)) {
             fa_warn(fa_string("SYSEX message to DLS, ignoring"));
         } else {
-            int status, data1, data2;
+            uint8_t status, data1, data2;
             fa_midi_message_decons(msg, &status, &data1, &data2);
             au_send_midi(context, status, data1, data2);
         }
